+++ Sources/SolanaSwift/Models/BufferLayout/SPLTokenMintState.swift
@@ -1,9 +1,9 @@
 import Foundation
 
-@available(*, deprecated, renamed: "SPLTokenState")
-public typealias Mint = SPLTokenState
+@available(*, deprecated, renamed: "SPLTokenMintState")
+public typealias Mint = SPLTokenMintState
 
-public struct SPLTokenState: BufferLayout, Equatable, Hashable, Encodable {
+public struct SPLTokenMintState: BufferLayout, Equatable, Hashable, Encodable {
     public static var BUFFER_LENGTH: UInt64 = 82
 
     public let mintAuthorityOption: UInt32
@@ -15,7 +15,7 @@ public struct SPLTokenState: BufferLayout, Equatable, Hashable, Encodable {
     public let freezeAuthority: PublicKey?
 }
 
-extension SPLTokenState: BorshCodable {
+extension SPLTokenMintState: BorshCodable {
     public init(from reader: inout BinaryReader) throws {
         mintAuthorityOption = try .init(from: &reader)
         mintAuthority = try? PublicKey(from: &reader)